mcp-servers/cats-mcp-server/pyproject.toml
[project]
name = "cats-mcp-server"
version = "0.1.0"
description = "FastMCP server for CATS (Complete Applicant Tracking System) API wrapper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp[all]>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
